1. Create Menu Brand  name空白沒有防呆(空白字元TRIM) 2. 刪除菜單時分類時，應有警告訊息，確定後才刪除。 3. 若刪除菜單分類，則此分類下的所有菜單皆會show不出來，     應該要把這些菜單原有的分類清空並一律歸類到"未分類"中 4.首頁ICON變小 5.新增或編輯菜單按到上一頁要提示是否要儲存 6.編輯菜單刪除後, 被刪除的MENU的分類要改為空白 7.Due Date一定要選擇才可以產生團購單

+++ app/src/main/java/com/iew/fun2order/ui/home/HomeFragment.kt
@@ -34,6 +34,7 @@ import com.iew.fun2order.R
 import com.iew.fun2order.db.dao.MenuTypeDAO
 import com.iew.fun2order.db.database.AppDatabase
 import com.iew.fun2order.db.firebase.USER_MENU
+import com.iew.fun2order.db.firebase.USER_MENU_INFORMATION
 import com.iew.fun2order.db.firebase.USER_PROFILE
 import com.iew.fun2order.nativead.TemplateView
 import com.iew.fun2order.ui.home.adapter.MenuItemAdapter
@@ -123,7 +124,7 @@ class HomeFragment : Fragment() {
         btnDeleteMenuType.setOnClickListener {
             var editTextMenuType = item.findViewById(R.id.editTextMenuType) as EditText
 
-            if (TextUtils.isEmpty(editTextMenuType.text))
+            if (TextUtils.isEmpty(editTextMenuType.text.trim()))
             {
                 editTextMenuType.requestFocus()
                 editTextMenuType.error = "類別不能為空白!"
@@ -145,7 +146,21 @@ class HomeFragment : Fragment() {
                  */
 
                 if(mUserProfile!!.brandCategoryList!!.size>0){
-                    deleteUserMenuFromFireBase(editTextMenuType.getText().toString(),item)
+
+                    //super.onBackPressed()
+                    androidx.appcompat.app.AlertDialog.Builder(requireContext())
+                        .setIcon(android.R.drawable.ic_dialog_alert)
+                        .setTitle("確認動作")
+                        .setMessage("確定刪除此類別[" + editTextMenuType.getText().toString() + "]?")
+                        .setPositiveButton(
+                            "確定"
+                        ) { dialog, which ->
+                            deleteUserMenuFromFireBase(editTextMenuType.getText().toString(),item)
+                            resetUserMenuFromFireBase(editTextMenuType.getText().toString())
+                             }
+                        .setNegativeButton("取消", null)
+                        .show()
+
 
                 }else{
                     editTextMenuType.requestFocus()
@@ -158,7 +173,7 @@ class HomeFragment : Fragment() {
             var editTextMenuType = item.findViewById(R.id.editTextMenuType) as EditText
 
 
-            if (TextUtils.isEmpty(editTextMenuType.text))
+            if (TextUtils.isEmpty(editTextMenuType.text.trim()))
             {
                 editTextMenuType.requestFocus()
                 editTextMenuType.error = "類別不能為空白!"
@@ -544,7 +559,7 @@ mMenuType=""
 
                    var test =  it.getValue(USER_MENU::class.java)
                     //if(menuType!=""){
-                        if(test!!.brandCategory==menuType){
+                        if(test!!.brandCategory.equals(menuType)){
 
                             mItemList.add(MenuItemListData(test!!.brandName, test!!.menuDescription, BitmapFactory.decodeResource(getResources(),R.drawable.image_default_member), test!!.menuImageURL ,test, mUserProfile))
                         }
@@ -708,6 +723,34 @@ mMenuType=""
 
     }
 
+    private fun resetUserMenuFromFireBase(menutype: String)
+    {
+        val context = this
+        val uuid =  FirebaseAuth.getInstance().currentUser!!.uid.toString()
+        var queryPath = "USER_MENU_INFORMATION/${uuid}/"
+        val myRef = Firebase.database.getReference(queryPath)
+        myRef.addListenerForSingleValueEvent(object : ValueEventListener {
+            override fun onDataChange(dataSnapshot: DataSnapshot) {
+
+                dataSnapshot.children.forEach()
+                {
+
+                    var test = it.getValue(USER_MENU::class.java)
+                    if(test!!.brandCategory.equals(menutype)){
+                        test!!.brandCategory=""
+                        //it.set (test)
+                        it.getRef().setValue(test)
+                    }
+                }
+
+            }
+            override fun onCancelled(error: DatabaseError) {
+
+            }
+        })
+
+    }
+
     private fun createMenuTypeButton(userProfile:USER_PROFILE)
     {
         mSegmentedGroupMenuType!!.removeAllViews()
@@ -722,7 +765,7 @@ mMenuType=""
             FrameLayout.LayoutParams.WRAP_CONTENT, FrameLayout.LayoutParams.WRAP_CONTENT
         )
 
-        if(mSegmentedGroupMenuType!!.childCount < 5) {
+        if(mSegmentedGroupMenuType!!.childCount < 4) {
             params.gravity = Gravity.CENTER
             mSegmentedGroupMenuType!!.layoutParams = params
         }else{
